Make cached_layer_write_metadata public
